refactor(dynamics365): consolidate shared helpers and constants
- Rename metadata files to match last subpackage segment (businesscentral.yaml, dataverse.yaml)
- Move parseAndThrowError and MAPPER from AbstractDataverseTask/AbstractBusinessCentralTask up to AbstractDynamics365Task
- Rename parseAndThrowODataError to parseAndThrowError at all dataverse call sites
- Extract ODATA_VERSION constant to AbstractDataverseTask, replacing "4.0" literals in Delete, Upsert, Query
