+++ src/main/resources/metadata/businesscentral.yaml
